Menu Items
Made an array of objects which is the menu items and displayed them on the page in a grid with their name and price and with a add button next to it. The button currently does not do anything but it will add the item to the cart. The actual rest of the menu items need to be added to the array.