Catering/Contact Us Page and fix ups
Allowed for information to be entered in and saved which will then get sent to the business. Once they press the submit button it comes up with an alert that lets the user know that their message has been sent to the business.

I also did the same thing with the contact us page so that the information that gets typed in is stored as variables and would then get sent to the business once they press the submit button. The actual process of the sending the information to the business is not made yet it but it will work in a way that the business will just get emailed with the information that is submitted.

For the menu page I have added in the rest of the menu items to the menu page with their prices and pictures and they all work fine with being added and removed from the cart.

On the your orde/cart page, once the user has pressed the submit button to submit their order, they get an alert to let them know that their order has been submitted, but now it doesn't output the user order after pressing the submit button. I don't think this is a huge issue for the MVP, because if really needed the user could also get an email with their order and their order details in a similar way to how the business would be receiving order information.