Fix ups and add ons
I added links to the business's social media pages in the footer and also fixed the footer. I added the logo next to the header and linked the order now button on the home page to the menu page and also fixed up some of the css on some of the other pages. Now I just need to code the part where the order information will get sent to the business in the form of an email