Email Testing
I did a test run of being able to send the emails. So I made a separate page for just testing the emails out where it would send the email to the business's email address and you can type in the subject of the email and the message. When you press the send button the email gets sent to the email address of the business and I have checked that the email it getting sent there and it does work. Now I just need to change the code so that the emails aren't on a separate page and will work on the pages needed as change it so that the messages and the subjects match the details of the user and their order and I can probably just do this by inserting the accordingly variables into those fields.