Emails
I fixed the email so now it displays all the items that the user has ordered and they all get sent in the email correctly with all the correct details. I also made so fix ups on the other pages, like fixing the fonts and putting in the about me paragraph.