More fix ups
Just some more fix ups with the css and content on the page. For the MVP I decided that it was fine was the contact us and catering pages to not have the forms on them and just display the contact info. I will incorporate this form feature in a later release.